Refactoring more file names, trying out tests using the Blazegraph database in Travis CI.
