+++ python_package/uploader/main.py
@@ -1,6 +1,6 @@
 __author__ = 'Stephen Kan'
 
-from ontology_uploader import upload_all_ontologies, create_namespace
+from blazegraph_upload import upload_all_ontologies, create_namespace
 from blazegraph_setup import generate_all
 import gc
 import _sparql
