+++ python_package/uploader/tests/test_metadata_validation.py
@@ -1,0 +1,1 @@
+__author__ = 'Stephen Kan'
