+++ python_package/uploader/tests/test_sequence_upload.py
@@ -1,0 +1,12 @@
+__author__ = 'Stephen Kan'
+
+import unittest
+from superphy.uploader.sequence_upload import SequenceUploader
+
+class SequenceUploaderTestCase(unittest.TestCase):
+
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
